+++ .vscode/settings.json
@@ -1,0 +1,4 @@
+{
+    "python.analysis.extraPaths": ["./src"],
+    "python.envFile": "${workspaceFolder}/.env"
+  }
